style: Clean up
Clean up directory structure and remove unwanted files
