+++ .gitignore
@@ -2,3 +2,4 @@ venv
 .idea
 out.json
 .DS_Store
+instance
