--- static/stylesheets/style.css
@@ -1,35 +1,0 @@
-body{
-    background-color:bisque
-}
-
-.search{
-    float:right;
-}
-
-#src{
-    width:200px;
-    height:30px;
-    font-size:15px;
-}
-
-#sort{
-    float:right;
-    width:100px;
-    height:100px;
-    font-size:15px;
-}
-
-#logo{
-    width:150px;
-    height:50px;
-}
-
-#recommend{
-    margin-top : 300px;
-    margin-left: 20px;
-}
-
-#navlist{
-    margin-bottom: 30px;
-}
-
