--- templates/404.html
@@ -1,7 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-
-<p><h1>Sorry, product unavailable</h1></p>
-
-{% endblock %}
