--- templates/Search.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <div>
-      <label for="Sort"></label>
-      <select name="Sort" id="Sort">
-        <option value="Price-asc"><a href="#link10">Price-Asc</a></option>
-        <option value="Price-dsc"><a href="#link11">Price-Dsc</a></option>
-        <option value="Relevance"><a href="#link12">Relevance</a></option>
-      </select>
-      
-{% endblock %}
