--- templates/base.html
@@ -1,41 +1,0 @@
-<html>
-<head>
-  <link rel="stylesheet" href="{{ url_for('static',filename='stylesheets/style.css') }}">
-  <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.js"></script>
-  <script type="text/javascript" src="{{ url_for('static', filename='/js/script.js') }}"></script>
-</head>
-<body>
-    <div id="navlist">
-        <a href="{{ url_for('index') }}"><img id="logo" src="{{ url_for('static',filename='img/bpp_fashions.png') }}"></a>
-        <div class="search">
-            <form action="submit">
-                <input id="src" type="text"
-                    placeholder=" Search "
-                    name="search">
-            </form>
-        </div>
-    </div>
-    <div id="Category">
-      <label for="category 1"></label>
-      <select name="category 1" id="category1">
-        <option value="S">None</option>
-        <option value="Men">Men</option>
-        <option value="Women">Women</option>
-        <option value="EXP">EXP</option>
-      </select>
-      <label for="category 2"></label>
-      <select name="category 2" id="category2">
-        <option value="None">None</option>
-        <option value="subcategory-1"><a href="#link4">Sub Category 1</a></option>
-        <option value="subcategory-2"><a href="#link5">Sub Category 2</a></option>
-        <option value="subcategory-3"><a href="#link6">Sub Category 3</a></option>
-      </select>
-      <button type="submit">Submit</button>
-    </div>
-  
-    {% block content %}
-
-    {% endblock %}
-
-</body>
-</html>
