--- templates/product.html
@@ -1,7 +1,0 @@
-{%extends "base.html"%}
-
-{% block content %}
-    <div id="recommend">
-      <p>Recommended For you</p>
-    </div>  
-{% endblock %}
